BG/Spells: Move battleground spell script to proper place

(based on wotlk/feature@2cb4be853)
