+++ src/game/AI/ScriptDevAI/scripts/world/spell_scripts.cpp
@@ -40,7 +40,6 @@ EndContentData */
 #include "Grids/GridNotifiers.h"
 #include "Grids/GridNotifiersImpl.h"
 #include "Grids/CellImpl.h"
-#include "OutdoorPvP/OutdoorPvP.h"
 
 /* When you make a spell effect:
 - always check spell id and effect index
@@ -445,54 +444,6 @@ struct SpellStackingRulesOverride : public SpellScript
     }
 };
 
-/*#####
-# spell_battleground_banner_trigger
-#
-# These are generic spells that handle player click on battleground banners; All spells are triggered by GO type 10
-# Contains following spells:
-# Arathi Basin: 23932, 23935, 23936, 23937, 23938
-# Alterac Valley: 24677
-# Isle of Conquest: 35092, 65825, 65826, 66686, 66687
-#####*/
-struct spell_battleground_banner_trigger : public SpellScript
-{
-    void OnEffectExecute(Spell* spell, SpellEffectIndex /*effIdx*/) const
-    {
-        // TODO: Fix when go casting is fixed
-        WorldObject* obj = spell->GetAffectiveCasterObject();
-
-        if (obj->IsGameObject() && spell->GetUnitTarget()->IsPlayer())
-        {
-            Player* player = static_cast<Player*>(spell->GetUnitTarget());
-            if (BattleGround* bg = player->GetBattleGround())
-                bg->HandlePlayerClickedOnFlag(player, static_cast<GameObject*>(obj));
-        }
-    }
-};
-
-/*#####
-# spell_outdoor_pvp_banner_trigger
-#
-# These are generic spells that handle player click on outdoor PvP banners; All spells are triggered by GO type 10
-# Contains following spells used in Zangarmarsh: 32433, 32438
-#####*/
-struct spell_outdoor_pvp_banner_trigger : public SpellScript
-{
-    void OnEffectExecute(Spell* spell, SpellEffectIndex /*effIdx*/) const
-    {
-        // TODO: Fix when go casting is fixed
-        WorldObject* obj = spell->GetAffectiveCasterObject();
-
-        if (obj->IsGameObject() && spell->GetUnitTarget()->IsPlayer())
-        {
-            Player* player = static_cast<Player*>(spell->GetUnitTarget());
-
-            if (OutdoorPvP* outdoorPvP = sOutdoorPvPMgr.GetScript(player->GetCachedZoneId()))
-                outdoorPvP->HandleGameObjectUse(player, static_cast<GameObject*>(obj));
-        }
-    }
-};
-
 struct GreaterInvisibilityMob : public AuraScript
 {
     void OnApply(Aura* aura, bool apply) const override
@@ -713,6 +664,4 @@ void AddSC_spell_scripts()
     RegisterSpellScript<RaiseDead>("spell_raise_dead");
     RegisterSpellScript<SplitDamage>("spell_split_damage");
     RegisterSpellScript<TKDive>("spell_tk_dive");
-    RegisterSpellScript<spell_battleground_banner_trigger>("spell_battleground_banner_trigger");
-    RegisterSpellScript<spell_outdoor_pvp_banner_trigger>("spell_outdoor_pvp_banner_trigger");
 }
